requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
torch==2.0.0
torchvision==0.15.0
Pillow==10.0.0
numpy==1.24.0
cv2-python==4.8.0.74
gunicorn==21.2.0
huggingface-hub==0.17.3
timm==0.9.7